+++ test/lib/thermite/cargo_test.rb
@@ -32,6 +32,24 @@ module Thermite
       mock_module.run_cargo_build('release')
     end
 
+    def test_inform_user_about_cargo_exception
+      _, err = capture_io do
+        assert_raises RuntimeError do
+          mock_module(optional_rust_extension: false).inform_user_about_cargo
+        end
+      end
+
+      assert_equal '', err
+    end
+
+    def test_inform_user_about_cargo_warning
+      _, err = capture_io do
+        mock_module(optional_rust_extension: true).inform_user_about_cargo
+      end
+
+      assert_equal mock_module.cargo_recommended_msg, err
+    end
+
     def described_class
       Tester
     end